move deprecated methods to a deprecated file
